requirements_exe.txt
Flask==2.3.3
openai==1.35.0
pyinstaller==5.13.2
requests==2.31.0
Pillow==10.0.0
urllib3==2.0.7
certifi==2023.7.22
